mqttpublisher/requirements.txt
paho-mqtt==1.5.0
requests==2.22.0
psutil==5.7.0